DoH: Make session resumption as a flag
Use a flag, doh_session_resumption, to control whether or not to
enable session resumption. If the flag is unset, session resumption
is disabled.

Because the value of the flag is cached in the DoH client, if
there are some networks existing before setting a new value to
the flag, those network will still use the old value of the flag
rather than the new one until private DNS settings changes.

Bug: 205922811
Test: cd packages/modules/DnsResolver && atest
Test: changed the flag and checked that DnsResolver behaved as
      expected
Change-Id: I6fb24d4251b6e0fc163d169bb9cc68703bb76812