Refactor Thread::SweepInterpreterCaches
Extracted from the shared interpreter cache CL.

Bring the loop-over-all-threads inside the sweep method,
so that we can handle shared cache there as well.

Also add DCHECK to assert that we handle all opcodes.
This will complain if we add opcode in the cache, but
forget to add sweep handler for it, or if the cache
consents gets corrupted/invalid for some reason.

Test: ./art/test.py -b -r --host
Change-Id: I491d9be898bbd35800645ba994907e4d9be83b80