hikey960/hikey: Re-transition to using to vendor-package
Moves the bootloader objects, and mali libraries to the vendor
package and adds the support files, tweaking the flash-all.sh
script

Will use the vendor-package located here:
  https://releases.linaro.org/android/aosp-linaro-vendor-package/extract-linaro_devices-20220210.tgz

Previously this patch was reverted, as there would be build
errors when the vendor pacakge was missing.

This patch re-applies the earlier change, with some additional
fixes to ensure we don't see build trouble when the package is
missing.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I53ec5516d91391c0d3f18c801e730eff150076af