Fix error handling when underlying network gets lost
This commit fixes the IKE library to fire onError callback and keep
the IKE Session alive when there is a network loss as long as
IKE_OPTION_MOBIKE is set. In this way, the caller still has an
opportunity to recover by setting a new network or waiting for a
new default network.

This commit changes the behaviour of reporting network loss back to
what it was before aosp/1881550 was submitted. Without this commit,
if the server does not support the RFC4555 MOBIKE, IKE Session will
be immediately killed even if the IKE Session can be recovered from
performing a Rekey-Child-based migration.

Bug: 220518120
Test: atest FrameworksIkeTests (new tests added), CtsIkeTestCases
Change-Id: Iccd01990397eaf13059ca4f6079e4506cc5917a7