process_symbols: Log informational messages to stdout, not stderr.
These are not error messages, so they should go to stdout.

It appears that the Python logging library is not capable of sending
different log levels to different streams, so just use print statements
directly instead of the logging library. Do the same for the existing
logging to stderr so that we can drop the dependency on the logging
library altogether.

Change-Id: Ifdafac657966a805aef1766e92182333c184af60