Add locally built bootstrap JDK
The OpenJDK17 build requires a 16 or 17 bootstrap JDK, which we don't
yet have prebuilts for.  Check in this locally built OpenJDK17, which
was itself bootstrapped against the debian openjdk-17-jdk package.
This will be replaced with a JDK built on the build servers once
they have completed a build bootstrapped from this JDK.

Bug: 219098645
Test: toolchain/jdk/build/build-openjdk17-linux.sh
Change-Id: I32377cc29db3b9b2f78f6a0a5bdbb6f6a7ccd3a3