Causing no error as removing routes on removed interface
In an IPv6 only network with clatd started, clatd will be stopped
first before destroying the network. The clat interface will
be removed from kernel while clatd stops, but the clat interface
was added with the network in the netd. Destroy the network will
iterate the adding interfaces and try to remove all of them. The
VPN local exclusion routes are added/removed with the existence
of the interfaces. It will cause an ENODEV error while trying to
find the interface index of the clat interface since the clat
interface was removed. The failure was considered as an error
and stop destroying the network. The routing will not be cleared
as expected and cause the left over rules.

Because stopping clat is always before destroying the network, it
is expected to get such error since the clat interface exists in
ConnectivityService if the network destroys with clat started.
Thus, stop considering this failure an error to finish all works
necessary for clearing interfaces in the network.

Sample issue rule:
16000:	from all fwmark 0x10063/0x1ffff iif lo lookup local_network
16000:	from all fwmark 0x10064/0x1ffff iif lo lookup wlan0
16000:	from all fwmark 0x10065/0x1ffff iif lo lookup wlan0
16000:	from all fwmark 0x10066/0x1ffff iif lo lookup wlan0
17000:	from all iif lo oif dummy0 lookup dummy0
17000:	from all iif lo oif wlan0 lookup wlan0
17000:	from all iif lo oif wlan0 lookup wlan0
17000:	from all iif lo oif wlan0 lookup wlan0
18000:	from all fwmark 0x0/0x10000 lookup legacy_system
19000:	from all fwmark 0x0/0x10000 lookup legacy_network
20000:	from all fwmark 0x0/0x10000 lookup local_network
23000:	from all fwmark 0x64/0x1ffff iif lo lookup wlan0
23000:	from all fwmark 0x65/0x1ffff iif lo lookup wlan0
23000:	from all fwmark 0x66/0x1ffff iif lo lookup wlan0

Bug: 184750836
Bug: 220997151

Test: cd system/netd ; atest
Test: Manually connect/disconnect IPv6 only WiFi and observe the
      routing is removed as expected
Change-Id: Ia981535a61dcc18aa25e7f35133c0fd822239b67