Autoupdate of 3 submodules
*external/error_prone updated to 7ef30d4799433c41f7d74794600deef5e503325d
Remove -Xep:EqualsNull:WARN

This was temporarily demonted from an error
to a warning. Now that all the occurances
of this bug have been removed, make it an
error again.

Bug: 190944875
Test: Presubmits
Change-Id: Ic47d531f9dafcf34a14dc141dfc776e02bdd4484

*external/jackson-databind updated to 814bd9b551e1f4c758c6b009404e9900942b263e
Demote EqualsNull error to warning

It is used intentially to check how a custom
subclass responds to null in this case.

Bug: 190944875
Test: Presubmits
Change-Id: If9cab7690230980a84a91e62ae30d0445ab8355f

*external/robolectric-shadows updated to b59877b659f1e1925938f58fafde71e40becc051
Merge "Demote EqualsNull error to warning"