Support non-core permissions
The @EnforcePermission annotation supports any permission defined in
frameworks/base/core/res/AndroidManifest.xml. Add support for
permissions declared outside this file.

If the permission name contains at least one dot ("."), it assumes that
it is a fully-qualified name and it is used as-is. Otherwise, it assumes
it is available within android.Manifest.permission.

In both cases, the generated code uses the Java constant (as opposed
to a string) to avoid referencing non-existing permission.

Bug: 220214993
Test: atest --host aidl_unittests
Change-Id: I268c1696ef65d36d5fa79e402d5094b63916beae