Correct the logic for updating mNetwork
Issue Description : When the device uses MOBIKE to do the Wifi to Wifi
handover, the net id has been changed but the IKE tunnel won't report
the tunnel disconnected so the resetTunnelManagerState() isn't called
(the mIsEpdgAddressSelected isn't cleared) and there is no chance to
update the new net id. If there is a new setup request coming at this
moment, the request will be processed with wrong net id.

Solution : From the code design, the IwlanDataService will iterate all
the tunnel state of the apns and call the updateNetwork to the
EpdgTunnelManager for the apn in TUNNEL_UP state to update the network
in the TunnelConfig. We update the mNetwork here to ensure the global
cache can be updated immediately.

Bug: 219401024
Test: On device test, VoWifi works
      atest IwlanTests
Change-Id: Id553f4de315f8879494ba115ddcd31d43d6de7a0