Add an faster OTA extractor
delta_generator is already capable of extracting images from an OTA.
However, it was created for testing purposes. Code in delta_generator
tries to mimic device environment as much as possible, and therefore
does not have optimal performance. Now go/bts-image-sharing project
needs an ota extractor in the cloud, re-write a faster alternative.

Test: ota_extractor ota.zip /tmp/extracted, 364s -> 272s, 25% faster.

Bug: 203712723
Bug: 221875271
Change-Id: Ie30375f2cd1df1890fc58a71f96b3aa430f0fd75