Autoupdate of 2 submodules
*packages/modules/Virtualization updated to a85f8595b2d81b4ca88e5df5ee7703716ae946d3
Remove run-as from Microdroid

It is only useful when there are apps present, which we don't have.

Bug: 223596375
Test: atest MicrodroidTests MicrodroidHostTestCases
Change-Id: Ifd0b8189a7f454fabf32ccfe0ae5a1b03c79c81a

*system/sepolicy updated to 7bde36e94eecdcb4f4788a0b657f71c838d007c9
Remove redundant sepolicy

We don't use MLS in Microdroid, so we don't need MLS rules, nor
mlstrusted[subject|object] labels. (We keep one MLS rule to satisfy
checkpolicy.)

A lot of attributes are unused in Microdroid, so we can remove their
declarations and any references to them. (That may not make the
compiled policy smaller, since hopefully they get optimised out
anyway, but it means there is less policy for humans to deal with.)

Remove labels that relate only to apps, which we don't have - MAC
permissions, run-as, seapp_contexts.

In passing, fix a comment snafu in both system & microdroid policy.

Bug: 223596375
Test: Run staged-apex-compile & compos_verify, no denials
Test: atest MicrodroidTests MicrodroidHostTestCases
Change-Id: Ifd3589945a2d8b4c0361e00eec5678795513fd8c