Optional parameters to take_bug_report
Changes test_name and begin_time parameters to be optional and defaults
to None. The test_name parameter is only used for logging, which is
redundant since this log will appear in the log file for the test. The
begin_time parameter is also used in logging, but is also used in the
filename for the resulting zip file. Using the beginning time for this
test is potentially dangerous if multiple bug reports are taken within a
single test. When this occurs, the zip file is overriden without warning
or error.

Bug: None
Test: NA
Change-Id: I7db5781ad6557d6441c8cff0bd58c90c51798ccb