Fix exit status detection for old devices.
Old devices don't have an adbd that supports shell_v2, which is required
for exit status propagation from the command run on the device to the
caller. As such, ExecutableTargetTest doesn't work on older devices.

To fix this, modify the command that is run to append the exit status
to stdout, and then parse that before returning the result to the
caller. This is only done for devices that don't have native support
for this behavior, and only when the "exit-status-workaround" device
option is set.

This resurrects aosp/1668906, which was rolled back in aosp/1708278,
with the following fixes and improvements:

1. Add bounds checking to fix the error from b/188056166.
2. Guard the exit status detection workaround with an option.
3. Add a unit test.

Test: Added unit test; ran run_tradefed_tests.sh
Change-Id: I5482264ee3fe122c3defd02f73040286cb774aa9