Scroll to the slice when navigating via flow arrows
To scroll to the selected slice, globals.sliceDetails need to be
up-to-date, which is not the case directly after changing the selection
because their computation requires a database call. This CL adds a new
state field with slice id to scroll to which is checked in
publishSliceDetails method.

A new parameter has been added to selectChromeSlice method, toggling
which to true will use the mechanism described above to scroll to the
slice after selection.

Bug: 224749275
Change-Id: I35f84fa9e7e02ec0389477c0b1bebd795d190cfe