pread() data into memory instead of lazily mmap everything
Initially, mmap() is used to ease parsing of manifest, as manifest has
variable size, and parsing it in a streaming fashoin is complicated.
Once manifest is parsed, applying ops in streaming manner is simple,
because each op has a defined data offset/length in manifest. Use
pread() to avoid loading the entire payload in memory at once.

Test: th
Bug: 221875271

Change-Id: I261c45bb81be096910e58ba5324410eb76dae1c9