Remove sdk_level dependency on libbase, libc++
Using sdk_level in a small (typically JNI) library that does not use
libbase or libc++ results in ~400kB of libraries to pull in, as it
depends on libbase and libc++.

Remove the dependency on libbase and libc++ by using NDK APIs and
C strings.

Test that a file that includes the headers can build by adding a
sdk_level_nosdk_static test library that is exercised in the unit test,
and does not build against any library. This ensures that new
dependencies are not added by mistake, as that test would fail to build.

Bug: 220819740
Test: atest unbounded_sdk_level_test
Change-Id: Iee0ca99dcdfdd87e3c2f53cfa313bb394ab31253