Remove usage of intermediates-dir-for
intermediates-dir-for relies on several variables that
haven't been set at the time the board configuration
runs. The board configuration got around that by using
deferred expansion, but deferred expansion is something
that starlark doesn't support. Remove intermediates-dir-for
by switching to TARGET_RECOVERY_FSTAB_GENRULE, which
accepts a genrule module to use instead of a file.

Bug: 201700692
Test: Presubmits
Change-Id: I6ebb75073abeebce71873f414129759789096b0f