BUILD: make demangle target a source_set, fix Wasm builds
static_library targets seem to create problems when building
Wasm on Mac. The only reason why the target is a static_library,
rather than a source_set, is for Bazel builds.
This CL adapts the target type according to the build system,
keeping it a source_set in standalone builds, and a static_library
when generating the Bazel BUILD files.

Bug: 204297290
Test: manual (builds on mac and UI works)
Change-Id: I077eb837cf423130d437b370a2fa9b0b0442ce6b