Autoupdate of 2 submodules
*device/generic/goldfish-opengl updated to 53e6ce168e992d753b9094430c8d1b3f097d6a73
Adds a NoOpComposer

... which is useful for bring-up of Cuttlefish on Gem5
which does not yet have virtio-gpu or a drm node. It
also does not have vsock which prevents the use of
Cuttlefish's "device config" for finding displays. In
that case, just have a default placeholder display that
we can configure later.

Also, clang-format.

Bug: b/226154708
Test: launch Cuttlefish in `noop` mode on Crosvm and
      check that device still reaches BOOT_COMPLETE
Change-Id: I258bf5b80b3f7523ff655aaf8d3e91c9cf76c047

*device/google/cuttlefish updated to 6ef951aabec308cd40bf1a52a36d29588f54e96a
Merge "Adds ro.vendor.hwcomposer.mode sysprop and set in Gem5"