Autoupdate of 2 submodules
*build/soong updated to 9e5cc8a6741e07cd9c433913016a9c8ce39f957e
Add pre-built attribute to host module definition for fake snapshot.

The host-snapshot provides host tools via prebuilt modules.
Having 2 prebuilts of the same module in a source tree leads to a build
time error.   Add new attribute when building the fake snapshot to
indicate that the host tool contains a prebuilt version.

When installing the host snapshot the user can opt not to include
modules that have a prebuilt version.

Bug: 225890931
Test: m HOST_FAKE_SNAPSHOT_ENABLE=true host-fake-snapshot dist
Change-Id: I044a92a280536f9c5ec93dcb277a5e5568cc4e42

*development updated to 3f04ab2fb4b33f0a6e1588a3ca5c29ab32b665d9
Merge "Do not install host snapshot modules with existing prebuilts."