Uniquely name ffx configs and logs
Change the location of ffx logs to the most precise location possible,
and include the time of creation in the target directory name. These
steps are necessary to differentiate between several instances of ffx
started within tests and prevent files from being overwritten when
multiple ffx isolated environments are created during a single test.
This happens, for example, when the device is rebooted. Fixes a bug
where multiple instances of ffx override each others' configs and logs.

Collpse the log directory into the ffx target directory since the log
filenames are self-documenting.

Bug: None
Test: act.py [...] -tc PolicyScanTest
Change-Id: I8b02c91947c7f49f1d442c3de3aee0b48374a091