pprof_builder: use trace_processor for demangling, and put obfuscated java frame names in system_name
This deduplicates code and takes advantage of the more powerful
demangling in trace processor.

One small side change is that if we see an obfuscated java method name
such as "A.b", and have deobfuscation packets with the full name, the
resulting pprof proto now has the human-readable name in Function.name,
and the obfuscated name in Function.system_name. This is the same as
what we do for mangled & demangled native frames. Previously the code
used the deobfuscated name for both name and system_name. Note that I
don't think any tools we use actually care about system_name, but I'm
retaining that info since it can still be investigated via the pprof
cmdline.

Bug: 204297290
Change-Id: Ic97cfbc98258d95c38df5a18981fb5b7de22bd52