Do not restore accpet_ra if the default IPv6 route has lost.
When detecting the loss of IPv6 default route the expected behavior
would be setting accept_ra to 0 and ignore the following RAs. But if
the device isn't in the link-local mode the accept_ra will be restored
to 2 immediately, which indicates to continue processing the RAs. That
isn't not aligned with the original design. Instead, pass the expected
accept_ra value when starting IPv6 stack in different scenarios.

This CL also updates the integration test and keep the test behavior
aligned with the expectation.

Bug: 226867453
Test: atest testIgnoreIpv6ProvisioningLoss_disableAcceptRa --iterations
Change-Id: I1ef3475d5772843825aba4397e4f7161542299c9