Suppress tidy checks that block migration to gnu11
The clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling
check gave more warnings in gnu11 mode and they are not suitable
for Android. This check is disabled in AOSP global default, and
local projects should not enable it or should change source code
to avoid those warnings.

Use tidy_checks_as_errors instead of tidy_flags to specify
the checks to be treated as errors.

Bug: 227361697
Test: make with global default set to gnu11
Change-Id: I09f7aa562ab51d3f536cd48184d430212d67b51b