dist: Also support declared directories.
When a target's default outputs include a declared directory
via ctx.actions.declare_directory, copy everything inside that
declared directory. Don't ignore directories.

The directory structure inside the declared directory is kept,
regardless of the --flat option. This is because the directory
itself is treated as an output, not its contents.

Test: bazel run //common:kernel_aarch64_abi_dist
Change-Id: Ib6bbd135cfb62be4df84b915bd2266b950e8f731