Autoupdate of 3 submodules
*external/adt-infra updated to 2d61a7211a37f56c0d94fceb583b0c464922d961
Use python 3 for build scripts when possible

This will start using python3 if the support is available. This will
allow us to migrate the build bots to python 3 without any interruption.

It will make use of virtual env to isolate our python dependencies.

Bug: 227774529
Change-Id: I2db060c248dc39d4eb529ef5588447f07a013b00

*external/qemu updated to da6f8de32bfc2e0dea1ad60643edd257c937b8ac
Merge "Python3 fixes for build scripts" into emu-master-dev
*tools/buildSrc updated to f51a9d4caad35d4ba6df1bcf0a09cad495e752aa
Python 3 fixes for build system

This makes sure that we can run the scripts under python3 and 2.

It will disable the qemu validation on python3. The qemu build scripts
do not work with python 3.

Bug: 227774529
Bug: 227780975
Change-Id: I565f1a0678a1efc5bcd527d4d3daa01cc85f4f8a