Fix using NULL as an integer
PostTask is passing NULL as a Message, which is being treated as 0
and implicitly constructing a Message.  It is valid to define NULL
as std::nullptr, which cannot be treated as 0 and fails to build.

This doesn't currently happen as this code is only compiled against
bionic, which uses Clang's definition of NULL as 0.  It does happen
in another copy of this code that is compiled for the host, which
when compiling against musl uses std::nullptr for NULL.

Bug: 218888599
Test: m wificond
Change-Id: I1f6e132ee68fa4f7c1480b250c0d6bd8a47e6baf