Fix NNAPI compilation time reported
TimeNanoMeasurer is a helper class that measures a time from the point
when it is initialized to the point when it is destroyed.

Prior to this change, the compilation time was wrong because it was
measured by a scoped TimeNanoMeasurer which was immediately destroyed.
This CL stores the TimeNanoMeasurer in a variable so that it will live
to the end of the function scope in order to correctly measure the time.

This change protects against these kinds of error in the future by
marking the TimeNanoMeasurer class as [[nodiscard]].

Finally, this change sets mTelemetryInfo to a default-initialized
TelemetryInfo object immediately before collecting timing data.

Bug: 228528606
Test: mma
Test: CtsNNAPITestCases (with local logging to verify the times)
Test: SANITIZE_HOST=address NeuralNetworksTest_static
Change-Id: I295b4ba05fdabd001fc1d827355e09d1de1b92a8