Remove unicast-response bit from .local resolution
RFC 6762 section 6.7 says that the Multicast DNS responder MUST send a
UDP response directly back to the querier, via unicast, to the query
packet's source IP address and port if the source UDP port in a received
Multicast DNS query is not 5353. Section 5.4 also mentions "...New
questions created by local clients afterwards should be treated as
normal 'QM' questions and SHOULD NOT have the unicast-response bit set
on the first question of the series."

DNS resolver works as a one-shot Multicast DNS querier, which send
queries from random ports. The unicast-response bit is not necessary to
be set.

Setting the unicast-response bit may also cause malfunction on .local
fallback queries. Currently, the bit is not cleared while .local
resolution is fallbacked from Multicast to Unicast DNS queries. DNS
server may send a no error response without Answer RR, or a failure.

One way to fix this is clearing the unicast-response bit before
fallback. However, it needs to parse packet bytes in res_nsend() because
the query packet is made before the res_nsend() but the fallback
decision is made in res_nsend(). Besides, it can cause problems in
cache. The query class is counted in cache key hashing. The answer
obtained by fallback queries is hashed (without unicast-response bit)
and stored in cache. But subsequent .local queries cannot match the
record because their unicast-response bit is initially set.

To avoid adding more complexity, the unnecessary unicast-response bit
should be removed.

Bug: 227147672
Test: atest
Test: Ping test.local on openWRT with fallback and without fallback.
Change-Id: Ib703a7537f638669fdc1d9c6927800e5c901786a