Allow @TestCaseName to be for names that are compatible with CTS/AJUR
Previously, @TestCaseName was disallowed altogether and the default
format was set to {method}[{index}] which was compatible with CTS/AJUR.
The intent was to prevent test authors from setting names which would
break CTS/AJUR features such as retry.

This change relaxes the restrictions to allow the use of @TestCaseName
as long as the generated test names were compatible with CTS/AJUR, i.e.
consisted of what could be a method name followed by an optional index
in brackets.

e.g. {method}_{0} would be allowed as long as the first parameter to
the test contained characters that are allowed in an identifier. An
invalid test name is detected as soon as it is created and will cause
the test run to fail.

Bug: 36541809
Test: Tried using @TestCaseName with valid and invalid patterns
Change-Id: I1dbc99b6b181898b438997f3e6b6d79a60b212b6