DoT quick fallback
If a DNS query that was sent to a DoT server times out, instead
of trying the next DoT server, let the query fall back to UDP.

This will be helpful when firewall blocks TCP/853 traffic, for
example captive portal re-login.

DoT quick fallback is enabled by default. It's configurable by
the flag `dot_quick_fallback`.

Bug: 228594312
Test: cd packages/modules/DnsResolver && atest
Test: On a wifi network with 2 validated DoT servers, blocked TCP/853
      traffic. DNS queries took 1x time w/ the flag enabled, and
      took 2x time w/o the flag enabled.
Change-Id: I27992baf4dd450391ab6241a5bf46b71ae656766