ftrace parsing: allow zero length __data_loc entries
__data_loc means that the kernel has stashed the string in the spare
bytes at the end of the ftrace page, and the value of the field encodes
the offset and the length of that string. See [1].

I strongly suspect this is the cause of the very rare crashes in
testlab/droidfood. We're overly strict with checking the string payload
offset, always assuming it's positive. Whereas it appears that in very
rare cases, both the offset and length could be zero.

[1] https://github.com/torvalds/linux/blob/e1f700ebd6bea293abe3c7e2807b252018efde01/include/trace/stages/stage5_get_offsets.h#L43-L50.
Bug: 204564312
Change-Id: Idcaee24100bfa6f57e3e3853a20b86e3e8f5a920