[DS#1] ResolverTest: assign specified socket type for GetAddrInfo_Dns64*
If the socket type is not specified, every address will appear
twice, once for SOCK_STREAM and one for SOCK_DGRAM. Just pick
one (SOCK_DGRAM in this commit) because the addresses for the
the second query of different socket types are responded to by
the cache. This helps to avoid that misunderstand the dns query
count of dns responder while people read the code.

Also:
1. Reword comments.
2. Simply the dns result checking via testing utility.
3. Use aggregate initialization on stuct addrinfo.

Bug: 159305929
Test: atest ResolverTest
Change-Id: I13a1a1c183b94cb921a01c4fc0cc6c7c7ba7227e