Autoupdate of 4 submodules
*cts updated to 9563b58259a6c053585dcf3e3f5bf501410c794f
Merge "Implement getInterfaceHash/Version"
*frameworks/base updated to 6f7ec05bd037dc1f5a4f890d5d200d194d058f14
Merge changes from topic "aidl-java-api-ci"

* changes:
  Implement getInterfaceHash/Version (omapi)
  Implement getInterfaceHash/Version (health)

*frameworks/opt/telephony updated to f03e4995cba4355f13774d2a3c33bb222688cd7d
Merge changes from topic "aidl-java-api-ci"

* changes:
  Implement getInterfaceHash/Version
  Implement getInterfaceHash/Version

*system/tools/aidl updated to 1f89532a35134490ff7dd610426f7d5cfc8aa797
Java: require interfaceVersion/hash for ToT vers.

This reqiures source changes, so we need to require it in order to
automate API finalization. The reasons for not having this for the
development version are mostly from the legacy C++ backend. For
minimalism here, the NDK and Rust backends are not considered.

Bug: 226954348
Test: build requires interface version and hash on interfaces.
Change-Id: I0b350d9cbc4aca62bff6dc268ee7419044fc6d01