Add support for dumping an entire BCC
Modify bcc::Chain::check() to return the verified Payload objects, and
expose the root public key from Chain.

Add implementations of Display to Payload and SubjectPublicKey to
allow them to be output in a human-readable form.

Add private struct ConfigDesc to represent the Configuration
Description from the payload, and implement Display for it.

Add a "--dump" option to "bcc_validator verify-chain" to print the
entire chain after successful verification.

Bug: 225177477
Test: Manual
Test: atest libcert_request_validator_tests
Change-Id: Id8985df6a7c7e8f1f5b220ceead87a24fa1b81ab