Autoupdate of 2 submodules
*build/soong updated to 8d0e466f190381e1cc9496c253ebbd8236debf79
Merge "Copy prebuilt objects to the intermediates directory"
*external/musl updated to 76db0349757e79ab0c702c73a171bea684bc5824
Add more files to the musl sysroot for rustc builds

Add libc++.a and the clang crtbegin and crtend objects to the musl
sysroot to allow building rustc against it.  The crtbegin and crtend
objects are also copied to all the different names gcc uses, as
clang uses the same object for all build types.

Bug: 216192129
Test: rustc builds
Change-Id: I681f90ee381405be50f32763d85e1f486fff142a