Fix ProguardMap line mapping
With R8 v3.1.4 the Proguard map stores line numbers slightly
differently.

Previously an entry having only one "clear" line of the form:
  <Start>:<End>:<Method info>:<OrigLine> -> <New name>
would map a line between Start and End to
  OrigLine + OffsetFromStart

In the new version, any line between Start and End gets mapped
precisely to OrigLine.

The ahat ProguardMap has been updated to check the version of the
map file and apply the correct mapping based on that.

Bug: 228000954
Test: m -j32 ahat ahat-tests && java -jar out/host/linux-x86/framework/ahat-tests.jar
Change-Id: I3a0ad58ccee1aae009d62759e59210a13530fc38