Add embedded_exec and exec_aspect.
exec_aspect tells a rule what the dependent target's
args and env are. Normally, they are only available
to `bazel run` and `bazel test` commands.

embedded_exec creates a script that runs the given
executable with its env and args. This removes
the env and args dependency so that we can
use $(rootpath) on the executable directly
to achieve the same effect of `bazel run`.

Test: manual
Test: bazel test //build/bazel_common_rules/exec/tests
  (see next CL)
Bug: 229350514
Bug: 228560650

Change-Id: I010e987618c31713d2018faf6b8dbe5310bcc1da