Autoupdate of 2 submodules
*external/libfuse updated to a57ca0355f00b44c8edfeccd6a094554a0731b69
ANDROID: fuse-bpf: Use fuse_bpf_args in uapi

fuse_args is not suitable for use in the uapi - it is not stable, and
contains internal pointers. Replace with stable equivalent.

Test: atest ScopedStorageDeviceTest passes
Bug: 202785178
Change-Id: I27d5ae0d7e0aaf30c51f3b97124737841605ae69

*system/bpfprogs updated to d4cf4d915b7ac33b5a6430da5f99a0860c0382d9
ANDROID: fuse-bpf: Use fuse_bpf_args in uapi

fuse_args is not suitable for use in the uapi - it is not stable, and
contains internal pointers. Replace with stable equivalent.

Test: atest ScopedStorageDeviceTest passes
Bug: 202785178
Change-Id: Id49ce2df8c1d137acb576074f1744884b65c7fc5