Autoupdate of 2 submodules
*external/sl4a updated to f1d2559b6698773a21a04c509c891073804b7bf5
Add BLE scan filter method for IRK in hex string format

Add a variation of bleSetScanFilterDeviceAddressTypeAndIrk method that
takes in an IRK in hex string format.

This approach was decided for simplicity for callers, since the
conversion from bytes->String in python and back to bytes in Java is
non-trivial (requires including charset to use, logic on either end to
translate the object - providing a method to handle conversion from hex
format in Java is much more straightforward than possible alternative).

Bug: 222349537
Test: gd/cert/run --device --sl4a_sl4a ConnectWithIrkTest
Change-Id: I5e5ab000d3c78f84bfd5411cee84a2f5661899c9

*packages/modules/Bluetooth updated to 7661959776f093d0327213742f41da00f1f7d2d0
Merge "Fix connect GATT with IRK and public address"