Autoupdate of 5 submodules
*device/google/bonito updated to b83708a0aa10c8601f1d2f1ed2dd75b52d97ed32
Use the ?= operator so BT properties can be overridden

System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: I29701f59e357d5896db5623493010be02dc2b4d5
Change-Id: I29701f59e357d5896db5623493010be02dc2b4d5

*device/google/coral updated to c3a9924696ca21a42c568788fdc259528c181d1b
Use the ?= operator so BT properties can be overridden

System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: I12dfdf4830e041b2c4e50eb8136ac0fb7146cc8e
Change-Id: I12dfdf4830e041b2c4e50eb8136ac0fb7146cc8e

*device/google/gs101 updated to e4a5a19b778a224646d5e619178071df36c8e9bf
Merge "Use the ?= operator so BT properties can be overridden"
*device/google/redbull updated to 0e1ada3c0103aa017f38c7d08c56e900bea2b8cb
Use the ?= operator so BT properties can be overridden

System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: Ie29700e8b87f6f3fc2da741411dfbe94fa56ca36
Change-Id: Ie29700e8b87f6f3fc2da741411dfbe94fa56ca36

*device/google/sunfish updated to e0a350461efb40bbbe116ecadf7cf3dc6cff84c6
Use the ?= operator so BT properties can be overridden

System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: Id6cd7901555edafef141bbc1bbd5e06dc3a54d0c
Change-Id: Id6cd7901555edafef141bbc1bbd5e06dc3a54d0c