Update TFLite build targets with AdServices tflite-support dependencies
Classes in tflite-support to be used in AdServices, including BertNLClassifier, depend on annotations, configurations, and op resolvers from the core tensorflow package.  However, these are not included in the current tensorflowlite_java build target, as it only tracks files from the top-level org/tensorflow/lite folder.  This CL updates the target to also include subfolders, and thereby the annotations, and creates new targets for the configurations and op resolvers.

Test: m
Bug: 228840235
Change-Id: Iedc9802a8f6086000cb3100354731df0fe0dc815