Autoupdate of 2 submodules
*hardware/interfaces updated to 53cf63d1fd1d4a4bd238560d93706bdc7c6363af
Merge "KeyMint VTS: x25519 pubkey as SubjectPublicKeyInfo"
*system/keymaster updated to 034b3247da4b3fd9908b5a4751f889a37386cacf
Expect x25519 data as SubjectPublicKeyInfo

Expect that the data for an X25519 key agreement operations should
arrive in the form of a SubjectPublicKeyInfo ASN1 structure rather than
just a raw key. This is more convenient and more consistent with the
expectations for other EC keys (using NIST curves).

Now that x25519 public keys are arriving in a SubjectPublicKeyInfo
envelope, make sure the correct error code is emitted when they
trigger a failure when erroneously attempting to perform key agreement
against a NIST curve private key.

Test: VtsAidlKeyMintTargetTest
Bug: 231959070
Change-Id: I0d515a7d3b7124288d36cbd219e3679fa5137f0c