Autoupdate of 2 submodules
*device/generic/goldfish-opengl updated to 03717120726e1071c6a7161cbcdb042a51e141a3
Merge "Pass VkExportFenceCreateInfo to host"
*device/generic/vulkan-cereal updated to 0492abedaad9169f3e1a91bf89cf6bf6685edb54
Create pool to reuse external fences instead of destroying.

When the guest creates a VkFence that is exportable to a SyncFd object,
it may be a valid operation for the guest to submit a vkDestroyFence
command while that VkFence is still unsignalled. Since this can result
in a race condition and deadlock on the host, we instead delay destruction
of these fences until device destruction. Once the fence has been signalled,
it can be reused in the place of another external fence.

Test: Ran emulator and tested anr freeze repro
Bug: b/228221208
Change-Id: I2b593ef569d96b454f1068440428826588ce78f5