Correct inode number type
The type of inode numbers used as directory names for dmabufs in sysfs
is unsigned long, but libdmabufinfo uses unsigned int for this type when
parsing. Until now this has been fine, since internally the kernel
actually only uses the range of an unsigned int, but this behavior has
recently changed so we should support the full range of an unsigned
long.

Bug: 232936707
Change-Id: Ia8dbbe8331904ac58101117ea1e0724a9100c179