Derive prerequisite target labels from build metadata
Hard-coded dependency target labels in test rules can break the Bazel
workspace if they are moved since the package path changes. Note that
moving a Soong build module to another path is not an issue since the
labels are unqualified and which doesn't impact modules that reference
them.

This change shields the workspace from these types of issues by mapping
prerequisite Soong build module names to their corresponding Bazel
target label using build metadata. The mapping is stored in a Starlark
file as variables which the test rules reference.

Bug: 232552492
Test: atest bazel_mode_test
Test: atest atest_unittests
Change-Id: I12d4cdb3e5135fe8d43007d567c44426a2afcc4d