Fix a corner case that module preparation retry pass but no tests are run.
There is a corner case that module retry is pass but no test cases are
run. In the current logic, TF reports preparation error at 1st run but
reports nothing becuase of no test results.

In such case, TF needs to report the results to notify it failed in the
1st run but passed afterwards after retrying.

Bug: 230423508
Bug: 231729160

Test: TreeHugger
Test: unittests

Change-Id: I8bd8bb58868df461ffd3e5e2ee9513a9e8d3804c