Move Java @EnforcePermission from interface to method
A permission annotation may be set on an AIDL interface or method. When
set on an interface, the corresponding Java annotation will be added to
the Java interface.

The intent of the Java annotation is to surface which permissions are
already verified by the aidl-generated service. When the Java annotation
is attached to the top-level interface, it is not visible to the
engineer editing/modifying the implementation method. If the subclass
also implements other methods, it is not clear which method is actually
covered by the permission checks.

This commit updates the generated Java annotation to always annotate the
method and not the interface.

Because an annotation cannot be duplicated on a method and annotations
may not easily be merged (e.g, anyOf={P1,P2} U anyOf={P3,P4}), force the
mutual exclusion of permission annotations on the AIDL interface or
method.

tl;dr:
- If an AIDL method is annotated for permission, the same annotation is
  added to the corresponding Java method.
- If an AIDL interface is annotated for permission, the annotation is
  added to all generated Java methods of this interface.
- If an AIDL interface is annotated for permission and a method is also
  annotated, raise an error.

Test: atest --host aidl_unittests
Bug: 231508256
Change-Id: Ie73eaad5ec197dba263b480c0ff0eb18444bf9ff