Infer data type by extension _with_ dot
FilePullerLogCollector tries to infer the LogDataType of a log file by
looking at the extension. When comparing the extension with a known
value, it is important to realize that the extension extracted from the
file contains the "." that separates the basename from the extension.

Two checks in FilePullerLogCollector did not account for that and were
therefor never recognized. This CL fixes that by including the dot in
the check.

Bug: 230070438
Test: presubmit
Change-Id: I3a47c09942fde265de1b58bb74cda524a5249e6c