Autoupdate of 2 submodules
*frameworks/native updated to 32a79d15620a28adfaf7f1dae26eb166440c5983
Merge "Reduce bug report compression level from 9 to 6."
*system/libziparchive updated to a46fc5be71ea936826eb8ca882eb41c7f501e446
Add default-compression flag to zip_writer

This change adds kDefaultCompression to zip_writer.h, which can be used
to set the compression level to gzip's default (6). Previously, the
compression level was always 9 (highest level of compression; slowest
speed). This is intended to allow for the compression level to be
decreased if speed is a consideration.

Test: Build
Bug: 215574756
Change-Id: Icd2fbcfa7ed11826cb819b5cdb1846174e9c15f2